// libs/community-query/src/config.d.ts
export declare const TAGS: {
    UPDATE_GROUP: string;
    GET_BENEFICIARY: string;
    GET_ALL_USER: string;
    GET_USER: string;
    GET_DASHBOARD: string;
    GET_ME: string;
    UPDATE_ME: string;
    UPDATE_USER: string;
    CREATE_USER: string;
    GET_ALL_PROJECTS: string;
    GET_ALL_ROLES: string;
    GET_ROLE: string;
    CREATE_ROLE: string;
    EDIT_ROLE: string;
    DELETE_ROLE: string;
    GET_ALL_CAMPAIGNS: string;
    GET_CAMPAIGNS: string;
    GET_CAMPAIGNS_AUDIO: string;
    GET_ALL_TRANSPORT: string;
    GET_ALL_AUDIENCE: string;
    GET_BENEFICIARIES: string;
    GET_BENEFICIARIES_STATUS: string;
    LIST_COMMUNITY_BENFICIARIES: string;
    CREATE_COMMUNITY_BENEFICARY: string;
    UPDATE_COMMUNITY_BENEFICARY: string;
    REMOVE_COMMUNITY_BENEFICARY: string;
    CREATE_BULK_COMMUNITY_BENEFICARY: string;
    CREATE_COMMUNITY_SETTINGS: string;
    LIST_COMMUNITY_SETTINGS: string;
    LIST_COMMUNITY_SETTINGS_NAME: string;
    UPDATE_COMMUNITY_SETTINGS: string;
    GET_VENDORS: string;
    GET_VENDOR_DETAILS: string;
    GET_PROJECT_DETAILS: string;
    LIST_COMMUNITY_GROUP: string;
    LIST_COMMUNITY_GROUP_BY_ID: string;
    ADD_COMMUNITY_GROUP: string;
    ADD_COMMUNITY_BENEFICIARY_GROUP: string;
    LIST_COMMUNITY_BENEFICIARY_GROUP: string;
    LIST_ACTIVE_FIELD_DEFINITIONS: string;
    LIST_COMMUNITY_FIELD_DEFINITIONS: string;
    CREATE_COMMUNITY_FIELD_DEFINITIONS: string;
    UPDATE_COMMUNITY_FIELD_DEFINITIONS: string;
    UPDATE_COMMUNITY_FIELD_DEFINITIONS_STATUS: string;
    DOWNLOAD_COMMUNITY_GROUPED_BENEFICIARIES: string;
    GET_COMMUNITY_BENEFICARY_STATS: string;
    LIST_TARGETING: string;
    ADD_BULK_FIELD_DEFINITIONS: string;
    CREATE_TARGETING: string;
    REMOVE_COMMUNITY_GROUP: string;
    PURGE_COMMUNITY_GROUP: string;
    GET_TARGETING_BENEFICIARIES: string;
    LIST_KOBO_SETTINGS: string;
    LIST_MAPPINGS_BY_IMPORT_ID: string;
    UPDATE_TARGETING_LABEL: string;
    DOWNLOAD_TARGETING_LABEL: string;
    LIST_PALIKA: string;
    EXPORT_TARGETED_BENEFICIARIES: string;
    VERIFY_BENEFICIARY: string;
    BULK_GENERATE_LINK: string;
};